requirements.txt
langchain
python-dotenv
faiss-cpu
huggingface_hub
langchain_community
langchain_huggingface
langchain_groq